[refactor] update admin.py
