+++ query/practice/add_sample_data2.sql
